Add oidcLdapUuidMatchingAnnotation sign in resolver
